ajustes RootReplace na tela home e perfil sair
