+++ Jeffrey/Jeffrey/Common/Extensions/ViewController+ReplaceRootViewController.swift
@@ -1,0 +1,24 @@
+//
+//  ViewController+ReplaceRootViewController.swift
+//  Jeffrey
+//
+//  Created by Michel dos Santos on 01/02/21.
+//
+
+import UIKit    
+
+
+extension UIViewController {
+    class func replaceRootViewController(viewController: UIViewController) {
+        guard let window = UIApplication.shared.windows.filter({ $0.isKeyWindow }).first
+        else {
+            return
+        }
+        let rootViewController = window.rootViewController!
+        viewController.view.frame = rootViewController.view.frame
+        viewController.view.layoutIfNeeded()
+        UIView.transition(with: window, duration: 0.3, options: .transitionFlipFromLeft, animations: {
+            window.rootViewController = viewController
+        }, completion: nil)
+    }
+}
