+++ Jeffrey/Jeffrey/Scenes/Home/View/HomeViewController.swift
@@ -29,7 +29,7 @@ class HomeViewController:  UIViewController, UITableViewDelegate, UITableViewDat
     // MARK: - LifeCircle
     override func viewDidLoad() {
         super.viewDidLoad()
-        self.navigationController?.popViewController(animated: true)
+//        self.navigationController?.popViewController(animated: true)
         viewModel = HomeViewModel()
         
         activity.showActivityIndicator(view: view, targetVC: self)
